move dockingManager to lib folder
#357
