+++ src/client/lib/dockingManager.js
