+++ src/client/src/ui/common/popout/openfinPopoutService.js
@@ -5,7 +5,7 @@ import { logger } from '../../../system';
 import OpenFinChrome from '../../common/components/openFinChrome/openFinChrome';
 import PopoutServiceBase from './popoutServiceBase';
 import _ from 'lodash';
-const DockingManager = require('exports?DockingManager!../../../system/openFin/dockingManager.js');
+const DockingManager = require('exports?DockingManager!../../../../lib/dockingManager.js');
 
 const DOCKED_CLASS_NAME = 'docked';
